refactor(e2e): relocate stranded multi_validator keystore utils
The e2e_multi_validator test was relocated to multi-node/block-production
on the consolidation PR, leaving src/e2e_multi_validator/utils.ts stranded.
Its sole consumer is the web3signer keystore test, so move it there and
drop the now-empty directory and its dangling bootstrap.sh references.

Part of A-1176
